Modified Post model and add view to filter newest on page
In the Post model, a field category, summary, image for summary and two fields for images for the main post was added.
In settings.py MEDIA_URL and MEDIA_ROOT for displaying images on the page added to the database was added.
In post_list.html, the loading of external files was changed to loading
from a static location.
In post_list.html the static articles to dynamic ones with a Django template variables and tags was changed.