Update README as well as Fixed first_visit site
In the views.py file, the display of the first_visit page for newly logged in users has been fixed.
In views.py removing unique code after user used was added.
Added new chapters on Video on YouTube and Deployment on Heroku in Readme.md.
In the forum_category_detail.html file, a working link has been added to the "ADD +" button.
In the calendar_follow.html file, changed the header on line 14.